requirements.txt
discord.py
aiosqlite
aiohttp
python-decouple